+++ ci-scripts/generateHtmlReport.py
@@ -1,25 +1,5 @@
 #!/usr/bin/env python3
-"""
-Licensed to the OpenAirInterface (OAI) Software Alliance under one or more
-contributor license agreements.  See the NOTICE file distributed with
-this work for additional information regarding copyright ownership.
-The OpenAirInterface Software Alliance licenses this file to You under
-the OAI Public License, Version 1.1  (the "License"); you may not use this file
-except in compliance with the License.
-You may obtain a copy of the License at
-
-  http://www.openairinterface.org/?page_id=698
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-------------------------------------------------------------------------------
-For more information about the OpenAirInterface (OAI) Software Alliance:
-  contact@openairinterface.org
----------------------------------------------------------------------
-"""
+# SPDX-License-Identifier: MIT
 
 #import argparse
 import os
@@ -58,7 +38,7 @@ class HtmlReport():
         with open(os.path.join(cwd, REPORT_NAME), 'w') as wfile:
             wfile.write(generate_header(args))
             wfile.write(generate_git_info(args))
-            wfile.write(build_summary(args, 'nef', '20', '8'))
+            wfile.write(build_summary(args, 'nef', '20', '9'))
             wfile.write(coding_formatting_log_check(args))
             wfile.write(analyze_sca_log_check())
             wfile.write(generate_footer())